+++ frontend/src/components/ChangeAutomationsDialog.jsx
@@ -43,7 +43,7 @@ function ChangeAutomationsDialog(props) {
 
   useEffect(() => {
     function fetchAutomations() {
-      fetch("/api/contrib/automations/" + automationType)
+      fetch("/api/automations/descriptors/" + automationType)
         .then((response) => {
             if (response.ok) {
               return response.json();